Predicate Syntax: Add desugaring and parse checks
This is (sort of) step 1/3 of implementing this. In this change, we have
implemented the desugaring of predicate functions into an empty function
and additional generated functions to typecheck the arguments of the
predicate.

The next steps will be the new `SpecChecker` and the actual encoding to
Viper pure functions. The SpecChecker checks that predicate functions
are never actually called and emits clear error messages for that case.
It could in the future have more tasks than just this.
The encoding to Viper will reuse large parts of what's already there for
`ProcedureSpecification`s, as those are encoded to pure functions
already and we'll duplicate a lot of code otherwise.